release v0.1.0b1: redis stream -> list
Keep it simple.  Don’t think we need a stream.

Significant simplification of the queue interface, including removal of receipt
handles.  The internal job queue doesn’t track message liveness: a get is a get
and immediately acks the message.  Liveness tracking is a job for a potential
top-level ingress queue, not for this inner queue which is more like a pubsub
anyway.